Move builtinNetworks into config
