--- extension/.env.example
@@ -1,4 +1,0 @@
-PRIVATE_KEY_STORAGE_KEY=default-private-key
-AGGREGATOR_URL=http://localhost:3000
-DEFAULT_CHAIN_ID=31337
-CRYPTO_COMPARE_API_KEY=
