--- extension/.env.release
@@ -1,4 +1,0 @@
-PRIVATE_KEY_STORAGE_KEY=quill-private-key
-AGGREGATOR_URL=https://arbitrum-testnet.blswallet.org
-DEFAULT_CHAIN_ID=421611
-CRYPTO_COMPARE_API_KEY=injected-from-github-actions-environment
